WIP: keep tile cache tests within LOC baseline (checkpoint)
